final commit

+++ Backend/src/main/java/ch/noseryoung/backend/domain/menu/MenuRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MenuRepository extends JpaRepository<MenuItem, Integer> {
+public interface MenuRepository extends JpaRepository<MenuItem, Long> {
 }
